LUCENE-2285: Code cleanups to remove compiler warnings in eclipse.
git-svn-id: https://svn.apache.org/repos/asf/lucene/java/trunk@917019 13f79535-47bb-0310-9956-ffa450edef68

+++ contrib/benchmark/src/java/org/apache/lucene/benchmark/byTask/tasks/NearRealtimeReaderTask.java
@@ -70,7 +70,7 @@ public class NearRealtimeReaderTask extends PerfTask {
     // Parent sequence sets stopNow
     reopenCount = 0;
     while(!stopNow) {
-      long waitForMsec = (long) (pauseMSec - (System.currentTimeMillis() - t));
+      long waitForMsec = (pauseMSec - (System.currentTimeMillis() - t));
       if (waitForMsec > 0) {
         Thread.sleep(waitForMsec);
         //System.out.println("NRT wait: " + waitForMsec + " msec");
