+++ src/java/org/apache/lucene/search/function/ShortFieldSource.java
@@ -39,10 +39,7 @@ import java.io.IOException;
  * composite (multi-segment) reader, this can easily cause
  * double RAM usage for the values in the FieldCache.  It's
  * best to switch your application to pass only atomic
- * (single segment) readers to this API.  Alternatively, for
- * a short-term fix, you could wrap your ValueSource using
- * {@link MultiValueSource}, which costs more CPU per lookup
- * but will not consume double the FieldCache RAM.</p>
+ * (single segment) readers to this API.</p>
  */
 public class ShortFieldSource extends FieldCacheSource {
   private FieldCache.ShortParser parser;
@@ -76,7 +73,7 @@ public class ShortFieldSource extends FieldCacheSource {
       /*(non-Javadoc) @see org.apache.lucene.search.function.DocValues#floatVal(int) */
       @Override
       public float floatVal(int doc) { 
-        return (float) arr[doc];
+        return arr[doc];
       }
       /*(non-Javadoc) @see org.apache.lucene.search.function.DocValues#intVal(int) */
       @Override
