+++ src/test/org/apache/lucene/util/TestIndexableBinaryStringTools.java
@@ -338,7 +338,7 @@ public class TestIndexableBinaryStringTools extends LuceneTestCase {
   public String binaryDump(byte[] binary, int numBytes) {
     StringBuilder buf = new StringBuilder();
     for (int byteNum = 0 ; byteNum < numBytes ; ++byteNum) {
-      String hex = Integer.toHexString((int)binary[byteNum] & 0xFF);
+      String hex = Integer.toHexString(binary[byteNum] & 0xFF);
       if (hex.length() == 1) {
         buf.append('0');
       }
@@ -359,7 +359,7 @@ public class TestIndexableBinaryStringTools extends LuceneTestCase {
   public String charArrayDump(char[] charArray, int numBytes) {
     StringBuilder buf = new StringBuilder();
     for (int charNum = 0 ; charNum < numBytes ; ++charNum) {
-      String hex = Integer.toHexString((int)charArray[charNum]);
+      String hex = Integer.toHexString(charArray[charNum]);
       for (int digit = 0 ; digit < 4 - hex.length() ; ++digit) {
         buf.append('0');
       }
